fix(admin) move into module
